# lumina-api/swagger.yaml
openapi: 3.0.3
info:
  title: Lumina Bot API
  version: 3.0.0
  contact:
    name: Lumina Dev Team
    url: https://bot.luminasink.com
  description: |
    # Lumina Bot API

    REST API for the Lumina Bot dashboard, inventory/gacha system, Discord guild
    management, and the internal bot ↔ API bridge. Built on Node.js/Express,
    deployed to Vercel (production) and `localhost:3000` (dev).

    ---

    ## Authentication

    The API supports **two** authentication methods for user-facing routes:

    1. **Cookie `lumina_token` (preferred for browsers)**
       - `httpOnly`, `Secure` (prod), `SameSite=Lax`, max-age = 1 hour
       - Set by `POST /expapi/v1/login` and the OAuth2 callback
       - Sent automatically by the browser — **immune to XSS** (JavaScript cannot read it)

    2. **Header `Authorization: Bearer <jwt>` (legacy / non-browser)**
       - For CLI tools, scripts, integration tests, and any client that does not
         manage cookies
       - Same JWT, same 1h expiry

    JWT expiry: **1 hour**. There is no refresh-token rotation yet — clients must
    re-authenticate when the token expires.

    The session helper `verifyRequestAuth(req)` reads the cookie first, then
    falls back to the Bearer header. Either is accepted on every protected route.

    ---

    ## CSRF Protection

    State-changing routes (`POST`/`PUT`/`DELETE` with `csrfProtectionNeeded: true`)
    require the **double-submit cookie** pattern:

...

...
    rotated hourly.

...

    ---

...

...

...

    ---
...
servers:
...
components:
  securitySchemes:

    cookieAuth:
      type: apiKey
      in: cookie
      name: lumina_token
      description: |
        httpOnly JWT cookie set by `POST /expapi/v1/login` or the OAuth2 callback.
        Preferred auth method for browsers (immune to XSS). Sent automatically by
        the browser; non-browser clients should use `bearerAuth` instead.

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <jwt>` — legacy/alternative auth for non-browser
        clients (CLI, scripts, integration tests). Same JWT, same 1h expiry as
        the cookie.

    internalKey:
      type: apiKey
      in: header
      name: internal-key
      description: |
        Required on `/expapi/internal/*` routes. Constant-time compared against
        `INTERNAL_API_KEY`. Only the Discord bot should possess this key.

    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Required on state-changing routes (`POST`/`PUT`/`DELETE`) that have
        `csrfProtectionNeeded: true`. The value must match the `csrf_token`
        cookie obtained from `GET /expapi/v1/csrf-token`.

    apiKeyQuery:
      type: apiKey
      in: query
      name: apiKey
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message (PT-BR).
        code:
          type: string
          description: Stable machine-readable error code.
          example: INVALID_CREDENTIALS
        details:
          type: object
          additionalProperties: true
          description: Optional structured details (e.g. attemptsRemaining, lockedUntil).
...
    Guild:
      type: object
      description: Discord guild managed by the bot.
      properties:
        guildId:
          type: string
          description: Discord guild ID (17–19 digits).
        guildReferenceName:
          type: string
        guildOwnerId:
          type: string
        memberCount:
          type: integer
        prefix:
          type: string
          default: l!
        guildLocale:
          type: string
          default: en-US
        memberWelcomeToggle:
          type: boolean
        memberDmToggle:
          type: boolean
        djEnabled:
          type: boolean
        persistentMute:
          type: boolean
        persistentWarns:
          type: boolean
        autoWarnPunishment:
          type: boolean
        canPunishStaff:
          type: boolean
        moderationChannelId:
          type: string
        memberJoinChannelId:
          type: string
        memberLeaveChannelId:
          type: string
        botInfoChannelId:
          type: string
        eventLogChannelId:
          type: string
        djRoleId:
          type: string
        muteRoleId:
          type: string
        banRoleId:
          type: string
        blockedChannels:
          type: array
          items:
            type: string
        warnsToMute:
          type: integer
        warnsToTimeOut:
          type: integer
        warnsToKick:
          type: integer
        warnsToBan:
          type: integer
        warnDuration:
          type: integer
        gachaMaxRolls:
          type: integer
        gachaGameMode:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    NewsPost:
      type: object
      description: News feed entry.
      properties:
        id:
          type: string
          description: MongoDB `_id` of the post.
        title:
          type: string
        body:
          type: string
          description: Full markdown body.
        excerpt:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        tag:
          type: string
          nullable: true
        pinned:
          type: boolean
        publishedAt:
          type: string
          format: date-time
        authorName:
          type: string
...
  responses:
    Unauthorized:
      description: Authentication missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Token inválido ou expirado.
            code: INVALID_TOKEN
...
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Erro interno do servidor.
            code: UNHANDLED_ERROR
paths:
...
  /expapi/v1/admin/guilds/{guildId}:
    put:
      tags: [Admin]
      summary: Update a guild's config (admin+)
      description: |
        Updates configuration fields on a managed guild. The set of writable
        fields depends on the caller's level:

        - Level ≥ 7: `djEnabled`, `memberWelcomeToggle`, `memberDmToggle`,
          `persistentMute`, `persistentWarns`, `autoWarnPunishment`,
          `canPunishStaff`
        - Level ≥ 8: `memberJoinChannelId`, `memberLeaveChannelId`,
          `moderationChannelId`, `botInfoChannelId`, `eventLogChannelId`,
          `djRoleId`, `muteRoleId`, `banRoleId`, `prefix`, `guildLocale`,
          `memberJoinMessage`, `memberLeaveMessage`, `memberJoinDmMessage`
        - Level ≥ 9: `blockedChannels`, `warnsToMute`, `warnsToTimeOut`,
          `warnsToKick`, `warnsToBan`, `warnDuration`, `gachaMaxRolls`,
          `gachaGameMode`

        CSRF-protected.
      operationId: updateGuild
      security:
        - bearerAuth: []
        - csrfToken: []
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
            pattern: '^\d{17,19}$'
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                djEnabled:
                  type: boolean
                memberWelcomeToggle:
                  type: boolean
                memberDmToggle:
                  type: boolean
                persistentMute:
                  type: boolean
                persistentWarns:
                  type: boolean
                autoWarnPunishment:
                  type: boolean
                canPunishStaff:
                  type: boolean
                memberJoinChannelId:
                  type: string
                memberLeaveChannelId:
                  type: string
                moderationChannelId:
                  type: string
                botInfoChannelId:
                  type: string
                eventLogChannelId:
                  type: string
                djRoleId:
                  type: string
                muteRoleId:
                  type: string
                banRoleId:
                  type: string
                prefix:
                  type: string
                guildLocale:
                  type: string
                memberJoinMessage:
                  type: string
                memberLeaveMessage:
                  type: string
                memberJoinDmMessage:
                  type: string
                blockedChannels:
                  type: array
                  items:
                    type: string
                warnsToMute:
                  type: integer
                warnsToTimeOut:
                  type: integer
                warnsToKick:
                  type: integer
                warnsToBan:
                  type: integer
                warnDuration:
                  type: integer
                gachaMaxRolls:
                  type: integer
                gachaGameMode:
                  type: string
            example:
              memberWelcomeToggle: true
              prefix: '!'
      responses:
        '200':
          description: Guild updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  updatedFields:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient permission or no allowed fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                permission:
                  value:
                    error: Permissão insuficiente.
                    code: INSUFFICIENT_PERMISSION
                no_fields:
                  value:
                    error: Nenhum campo permitido para alteração.
                    code: NO_ALLOWED_FIELDS
        '404':
          description: Guild not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Guilda não encontrada.
                code: GUILD_NOT_FOUND
        '500':
          $ref: '#/components/responses/ServerError'
...
  /expapi/v1/admin/news:
    post:
      tags: [Admin]
      summary: Create a news post (admin+)
      description: |
        Creates a new news post in the feed. Requires `accessType` level ≥ 7
        (admin+). The author name is derived from the authenticated admin's
        `firstName` + `lastName`.

        CSRF-protected.
      operationId: createNewsPost
      security:
        - bearerAuth: []
        - csrfToken: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title:
                  type: string
                body:
                  type: string
                  description: Full markdown body.
                excerpt:
                  type: string
                imageUrl:
                  type: string
                tag:
                  type: string
                pinned:
                  type: boolean
            example:
              title: New gacha season!
              body: |
                We just dropped 12 new skins in the Hextech chest pool...
              tag: gacha
              pinned: true
      responses:
        '201':
          description: Post created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    $ref: '#/components/schemas/NewsPost'
        '400':
          description: Missing `title` or `body`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: title e body são obrigatórios.
                code: MISSING_FIELDS
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient permission (level < 7).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Permissão insuficiente.
                code: INSUFFICIENT_PERMISSION
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Admin]
      summary: Delete a news post (admin+)
      description: |
        Deletes a news post by its MongoDB `_id`. Requires `accessType`
        level ≥ 7 (admin+).

        The `id` can be passed either as a query parameter (`?id=<mongoId>`)
        or in the request body (`{ id }`).

        CSRF-protected.
      operationId: deleteNewsPost
      security:
        - bearerAuth: []
        - csrfToken: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
        - in: query
          name: id
          required: false
          schema:
            type: string
          description: MongoDB `_id` of the post (alternative to body).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
      responses:
        '200':
          description: Post deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          description: Missing `id`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: id é obrigatório (query ou body).
                code: MISSING_ID
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient permission (level < 7).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Permissão insuficiente.
                code: INSUFFICIENT_PERMISSION
        '404':
          description: Post not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Post não encontrado.
                code: POST_NOT_FOUND
        '500':
          $ref: '#/components/responses/ServerError'

  # ───────────────────────────────────────────────────────────────────────
  # DEPRECATED
  # ───────────────────────────────────────────────────────────────────────
...
  /expapi/internal/updateguilddata:
    post:
      tags: [Internal]
      summary: Update a guild (whitelisted fields only)
      description: |
        Updates the guild document. Only fields in the bot whitelist are
        accepted; everything else is silently dropped (defense-in-depth against
        mass-assignment even if the internal key leaks).

        **Whitelisted fields:**
        `guildReferenceName`, `muteRoleId`, `banRoleId`, `moderationChannelId`,
        `memberCount`, `memberWelcomeToggle`, `memberJoinChannelId`,
        `memberJoinMessage`, `memberLeaveChannelId`, `memberLeaveMessage`,
        `memberJoinDmMessage`, `memberDmToggle`, `prefix`, `djEnabled`,
        `djRoleId`, `persistentMute`, `persistentWarns`, `warnsToMute`,
        `warnsToTimeOut`, `warnsToKick`, `warnsToBan`, `autoWarnPunishment`,
        `botInfoChannelId`, `eventLogChannelId`, `guildLocale`, `canPunishStaff`,
        `blockedChannels`, `gachaRolls`, `gachaMaxRolls`, `gachaGameMode`,
        `gachaRollsRefreshInterval`, `gachaRollsLastRefresh`.
      operationId: updateGuildData
      security:
        - internalKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [guildId]
              properties:
                guildId:
                  type: string
                guildReferenceName:
                  type: string
                prefix:
                  type: string
                guildLocale:
                  type: string
                memberCount:
                  type: integer
              additionalProperties: true
            example:
              guildId: '123456789012345678'
              prefix: 'l!'
              memberCount: 250
      responses:
        '200':
          description: Guild updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  updatedFields:
                    type: array
                    items:
                      type: string
                  updatedGuild:
                    $ref: '#/components/schemas/Guild'
        '400':
          description: Missing `guildId` or no valid whitelisted fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Nenhum campo válido para atualização.
                code: NO_VALID_FIELDS
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...